### 2026-02-08 #750 @martinezc1
Submission Inquiry: nycOpenData
Submitting Author Name: Christian Martinez
Submitting Author Github Handle: @martinezc1
Repository: https://github.com/martinezc1/nycOpenData
Version submitted: 0.1.6
Submission type: Standard

Version accepted: TBD

---

```
Package: nycOpenData
Title: Convenient Access to NYC Open Data API Endpoints
Version: 0.1.6
    c(
      person("Christian", "Martinez",
             email = "c.martinez0@outlook.com",
             comment = c(
               GitHub = "martinezc1",
               ORCID = "0009-0005-6026-6454"
             )),
      person("Crystal", "Adote",
             role = "ctb",
             comment = c(GitHub = "crystalna20")),
      person("Jonah", "Dratfield",
             role = "ctb",
             comment = c(GitHub = "jdratfield38")),
      person("Joyce", "Escatel-Flores",
             role = "ctb",
             comment = c(GitHub = "JoyceEscatel")),
      person("Rob", "Hutto",
             role = "ctb",
             comment = c(GitHub = "robhutto")),
      person("Isley", "Jean-Pierre",
             role = "ctb",
             comment = c(GitHub = "ijpier")),
      person("Shannon", "Joyce",
             role = "ctb",
             comment = c(GitHub = "shannonjoyce")),
      person("Laura", "Rose-Werner",
             role = "ctb",
             comment = c(GitHub = "laurarosewerner")),
      person("Emma", "Tupone",
             role = "ctb",
             comment = c(GitHub = "emmatup0205")),
      person("Xinru", "Wang",
             role = "ctb",
             comment = c(GitHub = "xrwangxr"))
    )
Description: Provides a unified set of helper functions to access datasets
    from the NYC Open Data platform <https://opendata.cityofnewyork.us/>.
    Functions return results as tidy tibbles and support optional filtering,
    sorting, and row limits via the Socrata API. The package includes endpoints
    for 311 service requests, DOB job applications, juvenile justice metrics,
    school safety, environmental data, event permitting, and additional citywide datasets.
    httr,
    curl
    scales,
    vcr (>= 0.6.0),
    webmockr

```

The package provides a programmatic interface to the Socrata-backed NYC Open Data API, replacing manual CSV downloads with reproducible R code.

Target Audience: The primary audience includes urban researchers, social scientists, and students in the CUNY system and beyond. It is specifically designed for analysts who need to integrate civic data into reproducible workflows without the overhead of manual data management.

Scientific Applications: nycOpenData facilitates large-scale urban studies, including spatial analysis of public service equity (via 311 data), longitudinal studies of urban ecology (via the Tree Census), and sociological research into school safety and juvenile justice trends. By providing a programmatic interface, it ensures that these urban analyses are fully reproducible and auditable.

The most closely related package is RSocrata, which provides a general-purpose interface for any Socrata-backed data portal. nycOpenData differs by meeting the "best-in-category" criteria through domain specialization.

Yes. The package exclusively accesses publicly available NYC Open Data datasets. It does not collect, infer, or augment personal data beyond what is provided by the data publisher, and it encourages responsible use of potentially sensitive fields (e.g., location data) in accordance with dataset-specific guidance and NYC Open Data terms.

https://github.com/ropensci/software-review/issues/748
@emilyriederer
@ronnyhdez

None. All pkgcheck checks pass locally and in continuous integration.
