### 2026-02-08 d911f4f
Add script, test, and vignette for NYC Shooting Incidents
