+++ vignettes/getting-started.Rmd
@@ -33,7 +33,7 @@ By using the `nyc_311()` function, we can gather the most recent 311 calls in Ne
 To start, let's pull a small sample to see what the data looks like. By default, the function pulls in the *10,000 most recent* requests, however, let's change that to only see the latest 3 requests. To do this, we can set `limit = 3`.
 
 ```{r small-sample}
-small_sample <- nyc_311(limit = 3)
+small_sample <- nyc_shooting_incidents(limit = 3)
 small_sample
 
 # Seeing what columns are in the dataset
