+++ man/nyc_shooting_incidents.Rd
@@ -1,0 +1,39 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/nyc_shooting_incidents.R
+\name{nyc_shooting_incidents}
+\alias{nyc_shooting_incidents}
+\title{[INSERT DATASET TITLE]}
+\source{
+NYC Open Data: <https://data.cityofnewyork.us/Public-Safety/NYPD-Shooting-Incident-Data-Year-To-Date-/5ucz-vwe8/about_data>
+}
+\usage{
+nyc_shooting_incidents(limit = 10000, filters = list(), timeout_sec = 30)
+}
+\arguments{
+\item{limit}{Number of rows to retrieve (default = 10,000).}
+
+\item{filters}{Optional list of field-value pairs to filter results.}
+
+\item{timeout_sec}{Request timeout in seconds (default = 30).}
+}
+\value{
+A tibble containing NYPD Shooting Incident Data (Year to Date) data.
+}
+\description{
+Downloads NYPD Shooting Incident Data (Year to Date) data from NYC Open Data.
+}
+\details{
+This is a data set that will show you every shooting incident that has occurred in New York City. It will show you the exact occur date, time, borough in which the shooting took place, location of the incident and much more. Researchers would use this dataset to find what time of day does the most shooting incidents take place or which location has the most shootings.
+}
+\examples{
+\donttest{
+if (curl::has_internet()) {
+  # Quick example (fetch 2 rows)
+  small_sample <- nyc_shooting_incidents(limit = 2)
+  small_sample
+
+  # Example with a filter (Change 'column_name' and 'value' to real fields)
+  # nyc_shooting_incidents(limit = 2, filters = list(column_name = "value"))
+}
+}
+}
