### 2026-02-08 245de38
Add script, test, and vignette for NYC Climate Projections: Extreme Events and Sea Level Rise
