### 2026-02-09 288db48
Add script, test, and vignette for NYC Wetlands
