### 2026-02-09 3d8171e
Add files via upload
Add NYCHA violations dataset, test, and vignette
