### 2026-02-10 03ff05c
Add script, test, and vignette for nyc_slash_stab
