+++ vignettes/investigating-inmate-incidents.Rmd
@@ -1,6 +1,6 @@
 ---
 title: "Investigating Inmate Incidents"
-output: rmardown::html_vignette
+output: rmarkdown::html_vignette
 author: "Isley Jean-Pierre"
 vignette: >
   %\VignetteIndexEntry{Investigating Inmate Incidents}
@@ -18,36 +18,30 @@ knitr::opts_chunk$set(
 ```{r setup, include=FALSE}
 knitr::opts_chunk$set(echo = TRUE)
 library(nycOpenData)
-library(tidyverse)
 library(ggplot2)
-library(jsonlite)
-library(httr)
 library(dplyr)
 ```
 
 ## Introduction
 
 This dataset contains information about the kind of violent incident that happens among inmates 
-in prison. There is row that provide the incidents' ID, the data they were reported, the kind
-of incident they were, and in what facilities those incidents occurred.A researcher might
-want to use this dataset to investigate the kind of crimes inmates commit while they are
-incarcerated.
+in prison. Each row provides the incident ID, the date the incident was reported, the incident type, and the facility where it occurred. A researcher might want to use this dataset to investigate the kind of crimes inmates commit while they are incarcerated.
 
-The `nycOpenData` package provides a streamlined interface for accessing New York City's vast open data resources. It connects directly to the NYC Open Data Portal. It is currently utilized as a primary tool for teaching data acquisition in [Reproducible Research Using R](https://martinezc1-textbook--reproducible-research-using-r.share.connect.posit.cloud), helping students bridge the gap between raw city APIs and tidy data analysis.
+The `nycOpenData` package provides a streamlined interface for accessing New York City's vast open data resources. It connects directly to the NYC Open Data Portal. It is currently utilized as a primary tool for teaching data acquisition in [Reproducible Research Using R](https://martinezc1-reproducible-research-using-r.share.connect.posit.cloud/), helping students bridge the gap between raw city APIs and tidy data analysis.
 
 By using the `nyc_slash_stab` function, we can investigate the most recent violent incidents that occurred among incarcerated individuals to understand their nature and fatality.
-
 
 ## Let's take a look at the dataset.
 
 To start, let's pull a small sample to see what the data looks like. By default, the function pulls in the *10,000 most recent* requests, however, let's change that to only see the latest 3 requests. To do this, we can set `limit = 3`.
 
 ```{r small-sample}
-small_sample <- nyc_slash_stab(limit = 3)
+small_sample <- nyc_slash_stab(limit = 100)
 small_sample
 
 # Seeing what columns are in the dataset
 colnames(small_sample)
+unique(small_sample$facility)
 ```
 
 
@@ -55,11 +49,11 @@ The `nyc_slash_stab()` function can filter based off any of the columns in the d
 
 ```{r filter-incident}
 
-incident_slash_stab <- nyc_slash_stab(limit = 3, filters = list(incident = "incident_type"))
+incident_slash_stab <- nyc_slash_stab(limit = 3, filters = list(incident_type = "Stabbing"))
 incident_slash_stab
 
 # Checking to see the filtering worked
-unique(incident_slash_stab$incident)
+unique(incident_slash_stab$incident_type)
 ```
 
 ## Note.
@@ -75,8 +69,8 @@ note of their severity. But for future inquiries:
 
 ```{r slashing and stabbing}
 # Creating the datasets
-slash <- nyc_slash_stab(limit = 50, filters = list(facility = "AMKC", incident = "slashing"))
-stab <- nyc_slash_stab(limit = 50, filters = list(facility = "AMKC", incident = "stabbing"))
[…]
