+++ DESCRIPTION
@@ -11,6 +11,9 @@ Authors@R:
                GitHub = "martinezc1",
                ORCID = "0009-0005-6026-6454"
              )),
+      person("Crystal", "Adote",
+             role = "ctb",
+             comment = c(GitHub = "crystalna20")),
       person("Jonah", "Dratfield", 
              role = "ctb",
              comment = c(GitHub = "jdratfield38")),
@@ -20,12 +23,12 @@ Authors@R:
       person("Rob", "Hutto", 
              role = "ctb",
              comment = c(GitHub = "robhutto")),
+      person("Isley", "Jean-Pierre",
+             role = "ctb",
+             comment = c(GitHub = "ijpier")),
       person("Shannon", "Joyce",
              role = "ctb",
              comment = c(GitHub = "shannonjoyce")),
-      person("Crystal", "Na",
-             role = "ctb",
-             comment = c(GitHub = "crystalna20")),
       person("Laura", "Rose-Werner",
              role = "ctb",
              comment = c(GitHub = "laurarosewerner")),
