+++ docs/reference/index.md
@@ -191,6 +191,8 @@ Other NYC Open Data wrappers
   : 2015-2018 School Closure Discharge Reporting
 - [`nyc_shooting_incidents()`](https://martinezc1.github.io/nycOpenData/reference/nyc_shooting_incidents.md)
   : NYPD Shooting Incident Data (Year To Date)
+- [`nyc_slash_stab()`](https://martinezc1.github.io/nycOpenData/reference/nyc_slash_stab.md)
+  : Inmate Incidents - Slashing and Stabbing
 - [`nyc_street_hail_livery_active()`](https://martinezc1.github.io/nycOpenData/reference/nyc_street_hail_livery_active.md)
   : Street Hail Livery (SHL) Drivers - Active
 - [`nyc_tlc_new_driver_app_status()`](https://martinezc1.github.io/nycOpenData/reference/nyc_tlc_new_driver_app_status.md)
