+++ docs/llms.txt
@@ -162,6 +162,8 @@ functions and documentation:
 - **Rob Hutto** ([@robhutto](https://github.com/robhutto)) — Added
   support for the **New York City Population by Borough, 1950–2040**
   dataset.
+- **Isley Jean-Pierre** ([@ijpier](https://github.com/ijpier)) — Added
+  support for **Inmate Incidents – Slashing and Stabbing**.
 - **Shannon Joyce** ([@shannonjoyce](https://github.com/shannonjoyce)) —
   Added support for the **NYC Wetlands** dataset, including tests and
   vignette.
@@ -375,6 +377,8 @@ Other NYC Open Data wrappers
   : 2015-2018 School Closure Discharge Reporting
 - [`nyc_shooting_incidents()`](https://martinezc1.github.io/nycOpenData/reference/nyc_shooting_incidents.md)
   : NYPD Shooting Incident Data (Year To Date)
+- [`nyc_slash_stab()`](https://martinezc1.github.io/nycOpenData/reference/nyc_slash_stab.md)
+  : Inmate Incidents - Slashing and Stabbing
 - [`nyc_street_hail_livery_active()`](https://martinezc1.github.io/nycOpenData/reference/nyc_street_hail_livery_active.md)
   : Street Hail Livery (SHL) Drivers - Active
 - [`nyc_tlc_new_driver_app_status()`](https://martinezc1.github.io/nycOpenData/reference/nyc_tlc_new_driver_app_status.md)
@@ -409,3 +413,5 @@ Other NYC Open Data wrappers
   Data](https://martinezc1.github.io/nycOpenData/articles/analyzing-nyc-shooting-incidents.md):
 - [Analyzing NYCHA Housing Maintenance Code
   Violations](https://martinezc1.github.io/nycOpenData/articles/nyc_ha_violations.md):
+- [Investigating Inmate
+  Incidents](https://martinezc1.github.io/nycOpenData/articles/investigating-inmate-incidents.md):
