+++ docs/articles/investigating-inmate-incidents.md
@@ -1,0 +1,181 @@
+# Investigating Inmate Incidents
+
+## Introduction
+
+This dataset contains information about the kind of violent incident
+that happens among inmates in prison. Each row provides the incident ID,
+the date the incident was reported, the incident type, and the facility
+where it occurred. A researcher might want to use this dataset to
+investigate the kind of crimes inmates commit while they are
+incarcerated.
+
+The `nycOpenData` package provides a streamlined interface for accessing
+New York City’s vast open data resources. It connects directly to the
+NYC Open Data Portal. It is currently utilized as a primary tool for
+teaching data acquisition in [Reproducible Research Using
+R](https://martinezc1-reproducible-research-using-r.share.connect.posit.cloud/),
+helping students bridge the gap between raw city APIs and tidy data
+analysis.
+
+By using the `nyc_slash_stab` function, we can investigate the most
+recent violent incidents that occurred among incarcerated individuals to
+understand their nature and fatality.
+
+## Let’s take a look at the dataset.
+
+To start, let’s pull a small sample to see what the data looks like. By
+default, the function pulls in the *10,000 most recent* requests,
+however, let’s change that to only see the latest 3 requests. To do
+this, we can set `limit = 3`.
+
+``` r
+small_sample <- nyc_slash_stab(limit = 100)
+small_sample
+#> # A tibble: 100 × 4
+#>    incident_id reported_dt             incident_type facility
+#>    <chr>       <chr>                   <chr>         <chr>   
+#>  1 227264      2026-01-24T21:10:00.000 Stabbing      GRVC    
+#>  2 226926      2026-01-17T19:58:00.000 Stabbing      OBCC    
+#>  3 226877      2026-01-16T17:51:00.000 Stabbing      NIC     
+#>  4 226700      2026-01-13T15:10:00.000 Stabbing      GRVC    
+#>  5 226465      2026-01-08T07:09:00.000 Stabbing      RNDC    
+#>  6 226408      2026-01-06T23:25:00.000 Stabbing      RMSC    
+#>  7 226398      2026-01-06T19:34:00.000 Stabbing      GRVC    
+#>  8 226206      2026-01-01T11:14:00.000 Stabbing      OBCC    
+#>  9 226110      2025-12-30T13:21:00.000 Stabbing      EMTC    
+#> 10 225938      2025-12-26T20:33:00.000 Stabbing      OBCC    
+#> # ℹ 90 more rows
+
+# Seeing what columns are in the dataset
+colnames(small_sample)
+#> [1] "incident_id"   "reported_dt"   "incident_type" "facility"
+unique(small_sample$facility)
+#> [1] "GRVC" "OBCC" "NIC"  "RNDC" "RMSC" "EMTC" "RESH"
+```
+
+The
+[`nyc_slash_stab()`](https://martinezc1.github.io/nycOpenData/reference/nyc_slash_stab.md)
+function can filter based off any of the columns in the dataset. To
+filter, we add `filters = list()` and put whatever filters we would like
+inside. From our `colnames` call before, we know that there is a column
+called “incident_type” which we can use to accomplish this.
+
+``` r
+
+incident_slash_stab <- nyc_slash_stab(limit = 3, filters = list(incident_type = "Stabbing"))
+incident_slash_stab
+#> # A tibble: 3 × 3
+#>   incident_id reported_dt             incident_type
+#>   <chr>       <chr>                   <chr>        
+#> 1 76348       2016-02-10T02:49:00.000 Stabbing     
+#> 2 78554       2016-04-23T12:54:00.000 Stabbing     
+#> 3 79040       2016-05-08T23:18:00.000 Stabbing
+
+# Checking to see the filtering worked
[…]

[…]
