### 2026-02-14 22524c0
Changed shooting incident data script and vignette
