+++ R/nyc_borough_population.R
@@ -23,10 +23,8 @@
 #'   small_sample <- nyc_borough_population(limit = 2)
 #'   small_sample
 #'
-#'   # Example with a filter (replace with a real field/value if desired)
 #'   nyc_borough_population(limit = 2, filters = list(borough = "Bronx"))
 #' }
-#'
 #' @export
 nyc_borough_population <- function(limit = 10000, filters = list(), timeout_sec = 30) {
   endpoint <- .nyc_endpoint("xywu-7bv9")
