+++ R/nyc_shooting_incidents.R
@@ -1,6 +1,6 @@
-#' NYPD Shooting Incident Data (Year To Date)
+#' NYPD Shootings
 #'
-#' Downloads NYPD Shooting Incident Data (Year to Date) data from NYC Open Data.
+#' Downloads NYPD Shooting Incident data from NYC Open Data.
 #'
 #' @param limit Number of rows to retrieve (default = 10,000).
 #' @param filters Optional list of field-value pairs to filter results.
@@ -8,9 +8,9 @@
 #' @return A tibble containing NYPD Shooting Incident Data (Year to Date) data.
 #'
 #' @details
-#' This is a data set that will show you every shooting incident that has occurred in New York City. It will show you the exact occur date, time, borough in which the shooting took place, location of the incident and much more. Researchers would use this dataset to find what time of day does the most shooting incidents take place or which location has the most shootings.
+#' This is a dataset identifies of every shooting incident that occurred in NYC since 2006.
 #'
-#' @source NYC Open Data: <https://data.cityofnewyork.us/Public-Safety/NYPD-Shooting-Incident-Data-Year-To-Date-/5ucz-vwe8/about_data>
+#' @source NYC Open Data: <https://data.cityofnewyork.us/Public-Safety/NYPD-Shootings/98wc-x49t/about_data>
 #'
 #' @examples
 #' # Examples that hit the live NYC Open Data API are guarded so CRAN checks
@@ -20,12 +20,11 @@
 #'   small_sample <- nyc_shooting_incidents(limit = 2)
 #'   small_sample
 #'
-#'   # Example with a filter (replace with real fields/values if desired)
-#'   # nyc_shooting_incidents(limit = 2, filters = list(boro = "BROOKLYN"))
+#'   nyc_shooting_incidents(limit = 2, filters = list(boro = "BROOKLYN"))
 #' }
 #' @export
 nyc_shooting_incidents <- function(limit = 10000, filters = list(), timeout_sec = 30) {
-  endpoint <- .nyc_endpoint("5ucz-vwe8")
+  endpoint <- .nyc_endpoint("98wc-x49t")
 
   query_list <- list(
     "$limit" = limit,
