+++ R/nyc_ha_violations.R
@@ -22,7 +22,6 @@
 #'   small_sample <- nyc_ha_violations(limit = 2)
 #'   small_sample
 #'
-#'   # Example with a filter (replace with a real field/value if desired)
 #'   nyc_ha_violations(limit = 2, filters = list(boro_nm = "BROOKLYN"))
 #' }
 #'
