+++ R/nyc_locallaw18_payreport.R
@@ -23,8 +23,7 @@
 #'   small_sample <- nyc_locallaw18_payreport(limit = 2)
 #'   small_sample
 #'
-#'   # Example with a filter (replace with real fields/values if desired)
-#'   # nyc_locallaw18_payreport(limit = 2, filters = list(race = "White"))
+#'   nyc_locallaw18_payreport(limit = 2, filters = list(race = "White"))
 #' }
 #' @export
 nyc_locallaw18_payreport <- function(limit = 10000, filters = list(), timeout_sec = 30) {
