+++ docs/articles/getting-started.md
@@ -41,34 +41,41 @@ however, let’s change that to only see the latest 3 requests. To do
 this, we can set `limit = 3`.
 
 ``` r
-small_sample <- nyc_shooting_incidents(limit = 3)
+small_sample <- nyc_311(limit = 3)
 small_sample
-#> # A tibble: 3 × 22
-#>   incident_key occur_date            occur_time boro  loc_of_occur_desc precinct
-#>   <chr>        <chr>                 <chr>      <chr> <chr>             <chr>   
-#> 1 313716449    2025-09-30T00:00:00.… 17:56:00   BRONX OUTSIDE           48      
-#> 2 313716449    2025-09-30T00:00:00.… 17:56:00   BRONX OUTSIDE           48      
-#> 3 313635117    2025-09-29T00:00:00.… 04:17:00   QUEE… OUTSIDE           102     
-#> # ℹ 16 more variables: jurisdiction_code <chr>, loc_classfctn_desc <chr>,
-#> #   location_desc <chr>, statistical_murder_flag <chr>, perp_age_group <chr>,
-#> #   perp_sex <chr>, perp_race <chr>, vic_age_group <chr>, vic_sex <chr>,
-#> #   vic_race <chr>, x_coord_cd <chr>, y_coord_cd <chr>, latitude <chr>,
-#> #   longitude <chr>, geocoded_column.type <chr>,
-#> #   geocoded_column.coordinates <list>
+#> # A tibble: 3 × 33
+#>   unique_key created_date           agency agency_name complaint_type descriptor
+#>   <chr>      <chr>                  <chr>  <chr>       <chr>          <chr>     
+#> 1 67976911   2026-02-13T02:05:17.0… NYPD   New York C… Illegal Parki… Blocked H…
+#> 2 67975725   2026-02-13T02:04:00.0… DOHMH  Department… Rodent         Condition…
+#> 3 67969953   2026-02-13T02:03:59.0… DOHMH  Department… Rodent         Rat Sight…
+#> # ℹ 27 more variables: location_type <chr>, incident_zip <chr>,
+#> #   incident_address <chr>, street_name <chr>, cross_street_1 <chr>,
+#> #   cross_street_2 <chr>, intersection_street_1 <chr>,
+#> #   intersection_street_2 <chr>, address_type <chr>, city <chr>,
+#> #   landmark <chr>, status <chr>, community_board <chr>,
+#> #   council_district <chr>, police_precinct <chr>, borough <chr>,
+#> #   x_coordinate_state_plane <chr>, y_coordinate_state_plane <chr>, …
 
 # Seeing what columns are in the dataset
 colnames(small_sample)
-#>  [1] "incident_key"                "occur_date"                 
-#>  [3] "occur_time"                  "boro"                       
-#>  [5] "loc_of_occur_desc"           "precinct"                   
-#>  [7] "jurisdiction_code"           "loc_classfctn_desc"         
-#>  [9] "location_desc"               "statistical_murder_flag"    
-#> [11] "perp_age_group"              "perp_sex"                   
-#> [13] "perp_race"                   "vic_age_group"              
-#> [15] "vic_sex"                     "vic_race"                   
-#> [17] "x_coord_cd"                  "y_coord_cd"                 
-#> [19] "latitude"                    "longitude"                  
-#> [21] "geocoded_column.type"        "geocoded_column.coordinates"
+#>  [1] "unique_key"               "created_date"            
+#>  [3] "agency"                   "agency_name"             
+#>  [5] "complaint_type"           "descriptor"              
[…]

[…]
