### 2026-02-19 2e6a854
Stop tracking pkgdown docs; build via GitHub Actions
