+++ docs/LICENSE-text.md
@@ -1,4 +1,0 @@
-# License
-
-    YEAR: 2025
-    COPYRIGHT HOLDER: Christian A. Martinez
