### 2026-02-19 c74efbb
Made significant changes to the package, including the helper function, each individual function, in relation to ropensci comments
