+++ R/nyc_dop_juvenile_rearrest_rate.R
@@ -26,15 +26,11 @@
 #' }
 #' @export
 nyc_dop_juvenile_rearrest_rate <- function(limit = 10000, filters = list(), timeout_sec = 30) {
-  endpoint <- .nyc_endpoint("c87b-2j3i")
-
-  query_list <- list(
-    "$limit" = limit,
-    "$order" = "year DESC"
+  .nyc_dataset_request(
+    dataset_id = "c87b-2j3i",
+    limit = limit,
+    filters = filters,
+    order = "year DESC",
+    timeout_sec = timeout_sec
   )
-
-  query_list <- .nyc_add_filters(query_list, filters)
-
-  data <- .nyc_get_json(endpoint, query_list, timeout_sec = timeout_sec)
-  tibble::as_tibble(data)
 }
