+++ R/nyc_street_hail_livery_active.R
@@ -28,15 +28,11 @@
 #' }
 #' @export
 nyc_street_hail_livery_active <- function(limit = 10000, filters = list(), timeout_sec = 30) {
-  endpoint <- .nyc_endpoint("5tub-eh45")
-
-  query_list <- list(
-    "$limit" = limit,
-    "$order" = "last_update_date DESC"
+  .nyc_dataset_request(
+    dataset_id = "5tub-eh45",
+    limit = limit,
+    filters = filters,
+    order = "last_update_date DESC",
+    timeout_sec = timeout_sec
   )
-
-  query_list <- .nyc_add_filters(query_list, filters)
-
-  data <- .nyc_get_json(endpoint, query_list, timeout_sec = timeout_sec)
-  tibble::as_tibble(data)
 }
