+++ R/nyc_daily_attendance_2018_2019.R
@@ -27,15 +27,11 @@
 #' }
 #' @export
 nyc_daily_attendance_2018_2019 <- function(limit = 10000, filters = list(), timeout_sec = 30) {
-  endpoint <- .nyc_endpoint("x3bb-kg5j")
-
-  query_list <- list(
-    "$limit" = limit,
-    "$order" = "date DESC"
+  .nyc_dataset_request(
+    dataset_id = "x3bb-kg5j",
+    limit = limit,
+    filters = filters,
+    order = "date DESC",
+    timeout_sec = timeout_sec
   )
-
-  query_list <- .nyc_add_filters(query_list, filters)
-
-  data <- .nyc_get_json(endpoint, query_list, timeout_sec = timeout_sec)
-  tibble::as_tibble(data)
 }
