+++ R/nyc_cause_of_death.R
@@ -25,15 +25,11 @@
 #' }
 #' @export
 nyc_cause_of_death <- function(limit = 10000, filters = list(), timeout_sec = 30) {
-  endpoint <- .nyc_endpoint("jb7j-dtam")
-
-  query_list <- list(
-    "$limit" = limit,
-    "$order" = "year DESC"
+  .nyc_dataset_request(
+    dataset_id = "jb7j-dtam",
+    limit = limit,
+    filters = filters,
+    order = "year DESC",
+    timeout_sec = timeout_sec
   )
-
-  query_list <- .nyc_add_filters(query_list, filters)
-
-  data <- .nyc_get_json(endpoint, query_list, timeout_sec = timeout_sec)
-  tibble::as_tibble(data)
 }
