+++ R/nyc_urban_park_animal_condition.R
@@ -28,16 +28,11 @@
 #'
 #' @export
 nyc_urban_park_animal_condition <- function(limit = 10000, filters = list(), timeout_sec = 30) {
-  endpoint <- .nyc_endpoint("fuhs-xmg2")
-
-  query_list <- list(
-    "$limit" = limit,
-    "$order" = "date_and_time_of_initial DESC"
+  .nyc_dataset_request(
+    dataset_id = "fuhs-xmg2",
+    limit = limit,
+    filters = filters,
+    order = "date_and_time_of_initial DESC",
+    timeout_sec = timeout_sec
   )
-
-  query_list <- .nyc_add_filters(query_list, filters)
-
-  data <- .nyc_get_json(endpoint, query_list, timeout_sec = timeout_sec)
-  tibble::as_tibble(data)
 }
-
