+++ R/nyc_period_attendance_reporting.R
@@ -26,15 +26,11 @@
 #' }
 #' @export
 nyc_period_attendance_reporting <- function(limit = 10000, filters = list(), timeout_sec = 30) {
-  endpoint <- .nyc_endpoint("hrsu-3w2q")
-
-  query_list <- list(
-    "$limit" = limit,
-    "$order" = "year DESC"
+  .nyc_dataset_request(
+    dataset_id = "hrsu-3w2q",
+    limit = limit,
+    filters = filters,
+    order = "year DESC",
+    timeout_sec = timeout_sec
   )
-
-  query_list <- .nyc_add_filters(query_list, filters)
-
-  data <- .nyc_get_json(endpoint, query_list, timeout_sec = timeout_sec)
-  tibble::as_tibble(data)
 }
