+++ R/nyc_pets_in_shelters.R
@@ -26,15 +26,12 @@
 #' }
 #' @export
 nyc_pets_in_shelters <- function(limit = 10000, filters = list(), timeout_sec = 30) {
-  endpoint <- .nyc_endpoint("5nux-zfmw")
-
-  query_list <- list(
-    "$limit" = limit,
-    "$order" = "date_year DESC"
+  .nyc_dataset_request(
+    dataset_id = "5nux-zfmw",
+    limit = limit,
+    filters = filters,
+    order = "date_year DESC",
+    timeout_sec = timeout_sec
   )
-
-  query_list <- .nyc_add_filters(query_list, filters)
-
-  data <- .nyc_get_json(endpoint, query_list, timeout_sec = timeout_sec)
-  tibble::as_tibble(data)
 }
+

[…]
