### 2026-02-19 f6f6ea2
Made changes to utils_request.R to account for comment 16 and 17 on the review, and also changed the borough population vignette to reflect the clean_names() change
