+++ tests/testthat/fixtures/nyc_311_2010_2019_test.yml
@@ -1,30 +1,0 @@
-http_interactions:
-- request:
-    method: GET
-    uri: https://data.cityofnewyork.us/resource/76ig-c548.json?%24limit=2&%24order=created_date+DESC
-  response:
-    status: 200
-    headers:
-      server: nginx
-      date: Thu, 19 Feb 2026 20:49:47 GMT
-      content-type: application/json;charset=utf-8
-      transfer-encoding: chunked
-      connection: keep-alive
-      access-control-allow-origin: '*'
-      etag: '"aW5kaWEuMTYzMTc0XzJfNTYyOG9HVUdSV2w3WER3WGI0THAzc3FIZE1DVnhvZw---gzipmGvjAYS6nnA_YHS9hlRIqwZ1jB--gzip--gzip"'
-      x-soda2-fields: '["unique_key","created_date","closed_date","agency","agency_name","complaint_type","descriptor","descriptor_2","location_type","incident_zip","incident_address","street_name","cross_street_1","cross_street_2","intersection_street_1","intersection_street_2","address_type","city","landmark","facility_type","status","due_date","resolution_description","resolution_action_updated_date","community_board","council_district","police_precinct","bbl","borough","x_coordinate_state_plane","y_coordinate_state_plane","open_data_channel_type","park_facility_name","park_borough","vehicle_type","taxi_company_borough","taxi_pick_up_location","bridge_highway_name","bridge_highway_direction","road_ramp","bridge_highway_segment","latitude","longitude","location"]'
-      x-soda2-types: '["text","floating_timestamp","floating_timestamp","text","text","text","text","text","text","text","text","text","text","text","text","text","text","text","text","text","text","floating_timestamp","text","floating_timestamp","text","text","text","text","text","number","number","text","text","text","text","text","text","text","text","text","text","number","number","point"]'
-      x-soda2-data-out-of-date: 'false'
-      x-soda2-truth-last-modified: Wed, 24 Dec 2025 20:34:02 GMT
-      x-soda2-secondary-last-modified: Wed, 24 Dec 2025 20:34:02 GMT
-      last-modified: Wed, 24 Dec 2025 20:34:02 GMT
-      vary: Accept-Encoding
-      content-encoding: gzip
-      strict-transport-security: max-age=31536000; includeSubDomains
-      x-socrata-requestid: 27088db14602eb6a44bc928926741300
-    body:
-      string: |
[…]

[…]
