+++ .github/workflows/pkgdown.yaml
@@ -13,12 +13,24 @@ jobs:
     runs-on: ubuntu-latest
 
     steps:
-      - uses: actions/checkout@v4
-      - uses: r-lib/actions/setup-r@v2
-      - uses: r-lib/actions/setup-r-dependencies@v2
+      - name: Checkout repository
+        uses: actions/checkout@v4
+
+      - name: Setup R
+        uses: r-lib/actions/setup-r@v2
+
+      - name: Install dependencies
+        uses: r-lib/actions/setup-r-dependencies@v2
         with:
           extra-packages: pkgdown
-      - name: Build site
+
+      - name: Build pkgdown site
         run: R -e 'pkgdown::build_site(override = list(destination = "docs"))'
-      - name: Deploy site to gh-pages
+
+      - name: Deploy to gh-pages
         uses: peaceiris/actions-gh-pages@v3
+        with:
+          github_token: ${{ secrets.GITHUB_TOKEN }}
+          publish_dir: docs
+          publish_branch: gh-pages
+          force_orphan: true
