### 2026-02-20 #750 @ronnyhdez
Hi @martinezc1!

Thanks for going through all the suggestions! I'm really glad to see the feedback was helpful!

One small process note that will help us move forward smoothly: could you please reply referencing each numbered comment (r02, r03, r05, etc.) and briefly indicate how it was addressed?

For example, I expect to see something like:

- r05: Updated the GitHub Actions workflow to trigger on main (see commit abc123).
- r07: Refactored repeated logic into a new internal function `.query_dataset()`.
- r16: I was 50/50 on r16 as I originally wanted to keep it matching the API definitions, but think using janitor::clean_names() (which I did) makes it better, especially for someone who may be new to R.

It doesn’t need to be super long. The goal is just to keep the review traceable and make sure we don’t miss any required items from me, or for the next stage, from the reviewers.

Thanks again! Is great to see the work you are putting in the package!
