### 2026-03-04 c6b0cf2
Made major changes to the package, reducing the individual functions and turning it into 3 functions- list, any, and pull
