+++ R/nyc_borough_community_report.R
@@ -1,36 +1,0 @@
-#' NYC Borough/Community District Report
-#'
-#' Downloads Borough/Community District Report data from NYC Open Data.
-#'
-#' @param limit Number of rows to retrieve (default = 10,000).
-#' @param filters Optional list of field-value pairs to filter results.
-#' @param timeout_sec Request timeout in seconds (default = 30).
-#' @return A tibble containing Borough/Community District Report data.
-#'
-#' @details
-#' This dataset provides monthly reports by borough and community district,
-#' covering metrics such as service requests, inspections, and community outreach.
-#' It’s published by NYC Department of Social Services for transparency and analysis.
-#'
-#' @source NYC Open Data: <https://data.cityofnewyork.us/resource/5awp-wfkt>
-#'
-#' @examples
-#' # Examples that hit the live NYC Open Data API are guarded so CRAN checks
-#' # do not fail when the network is unavailable or slow.
-#' if (interactive() && curl::has_internet()) {
-#'   # Quick example (fetch 2 rows)
-#'   small_sample <- nyc_borough_community_report(limit = 2)
-#'   small_sample
-#'
-#'   nyc_borough_community_report(limit = 2, filters = list(borough = "Bronx"))
-#' }
-#' @export
-nyc_borough_community_report <- function(limit = 10000, filters = list(), timeout_sec = 30) {
-  .nyc_dataset_request(
-    dataset_id = "5awp-wfkt",
-    limit = limit,
-    filters = filters,
-    order = "month DESC",
-    timeout_sec = timeout_sec
-  )
-}
