### 2026-03-04 9f4cce0
Made change to references section of pkgdown.yml
