+++ articles/Analyzing-the-Local-Law-97-of-2021-Pets-in-Shelter-Report.md
@@ -1,0 +1,216 @@
+# Analyzing the Local Law 97 of 2021-Pets in Shelter Report
+
+## Introduction
+
+New York City serves a large and diverse population, including
+individuals and families who experience housing instability and may rely
+on homeless shelters for temporary support. Pets serve many important
+roles in people’s lives, and understanding pet ownership among shelter
+entrants helps inform shelter policy, access to services, and resource
+planning.
+
+The **Local Law 97 of 2021 – Pets in Shelter Report** provides quarterly
+survey data on the types of pets owned by people entering New York City
+homeless shelters. This dataset was recently added to NYC Open Data and
+made publicly available on October 14, 2025. These data are published
+through NYC Open Data and are available
+[here](https://data.cityofnewyork.us/Social-Services/Local-Law-97-of-2021-Pets-in-Shelter-Report/5nux-zfmw/about_data).
+
+The `nycOpenData` package provides a streamlined interface for accessing
+New York City’s open data resources directly in R. By using the
+`nyc_locallaw97_shelter_pets()` function, we can retrieve the most
+recent quarterly shelter pet data and explore trends in pet ownership
+across time using familiar data analysis tools.
+
+## Pulling a Small Sample
+
+To start, we’ll pull a small sample of the data to get a sense of what
+it looks like. By default, the function retrieves the 10,000 most recent
+requests, so we’ll adjust this and only load the latest 3 by using
+`limit = 3`.
+
+``` r
+small_sample <- nyc_pull_dataset("nyc_locallaw97_shelter_pets", limit = 3)
+
+# Seeing what columns are in the dataset
+names(small_sample)
+#>  [1] "date_year"                      "date_quarter"                  
+#>  [3] "total_applicants"               "had_pet"                       
+#>  [5] "client_borough_of_entry_bronx"  "client_borough_of_entry"       
+#>  [7] "client_borough_of_entry_1"      "clients_who_would_have_entered"
+#>  [9] "clients_choosing_to_forego"     "chose_to_forego_shelter_for"   
+#> [11] "clients_who_intend_to_regain"   "number_of_dogs"                
+#> [13] "number_of_cats"                 "number_of_birds"               
+#> [15] "number_of_small_mammals"        "number_of_reptiles_small"      
+#> [17] "number_of_fish"                 "did_not_report_pet_type"       
+#> [19] "pet_s_placed_with_foster"       "pet_s_surrendered_to_animal"   
+#> [21] "pet_s_placed_with_family"       "pet_s_placed_with_friend"      
+#> [23] "pet_s_placed_with_other"        "pet_s_placed_with_non_profit"
+```
+
+Fantastic! We successfully pulled the
+`Local Law 97 of 2021-Pets in Shelter Report` data from the NYC Open
+Data Portal using the `nyc_locallaw97_shelter_pets()` function.
+
+Let’s now look at an example using the most recent 3 quarters of data.
+The `nyc_locallaw97_shelter_pets()` function allows users to filter
+based on any of the columns in the dataset. To apply filters, we add
+`filters = list()` and specify the desired conditions. From our earlier
+inspection of the column names, we know that there are columns such as
+`date_year` and `date_quarter` that can be used to filter the data by
+time period.
+
+``` r
+
+recent_quarters <- nyc_pull_dataset(
+  "nyc_locallaw97_shelter_pets", 
+  limit = 3,
+  filters = list(date_year = 2024))
+
+recent_quarters
+#> # A tibble: 3 × 24
[…]
