+++ articles/investigating-inmate-incidents.md
@@ -1,0 +1,196 @@
+# Investigating Inmate Incidents
+
+## Introduction
+
+This dataset contains information regarding the [Inmate Incidents -
+Slashing and
+Stabbing](https://data.cityofnewyork.us/Public-Safety/Inmate-Incidents-Slashing-and-Stabbing/gakf-suji/about_data)
+dataset on the NYC Open Data portal. Each row provides the incident ID,
+the date the incident was reported, the incident type, and the facility
+where it occurred. A researcher might want to use this dataset to
+investigate the kind of crimes inmates commit while they are
+incarcerated.
+
+The `nycOpenData` package provides a streamlined interface for accessing
+New York City’s vast open data resources. It connects directly to the
+NYC Open Data Portal. It is currently utilized as a primary tool for
+teaching data acquisition in [Reproducible Research Using
+R](https://martinezc1-reproducible-research-using-r.share.connect.posit.cloud/),
+helping students bridge the gap between raw city APIs and tidy data
+analysis.
+
+By using the `nyc_slash_stab()` function, we can investigate the most
+recent violent incidents that occurred among incarcerated individuals to
+understand their nature and fatality.
+
+## Let’s take a look at the dataset.
+
+To start, let’s pull a small sample to see what the data looks like. By
+default, the function pulls in the *10,000 most recent* requests,
+however, let’s change that to only see the latest 3 requests. To do
+this, we can set `limit = 3`.
+
+``` r
+small_sample <- nyc_pull_dataset("nyc_slash_stab", limit = 3)
+small_sample
+#> # A tibble: 3 × 4
+#>   incident_id reported_dt         incident_type facility
+#>         <dbl> <dttm>              <chr>         <chr>   
+#> 1      227264 2026-01-24 21:10:00 Stabbing      GRVC    
+#> 2      228684 2026-02-27 22:05:00 Stabbing      RNDC    
+#> 3      228558 2026-02-25 15:43:00 Stabbing      OBCC
+
+# Seeing what columns are in the dataset
+names(small_sample)
+#> [1] "incident_id"   "reported_dt"   "incident_type" "facility"
+
+small_sample |>
+  distinct(facility)
+#> # A tibble: 3 × 1
+#>   facility
+#>   <chr>   
+#> 1 GRVC    
+#> 2 RNDC    
+#> 3 OBCC
+```
+
+The `nyc_slash_stab()` function can filter based on any of the columns
+in the dataset. To filter, we add `filters = list()` and put whatever
+filters we would like inside. From our
+[`names()`](https://rdrr.io/r/base/names.html) call before, we know that
+there is a column called “incident_type” which we can use to accomplish
+this.
+
+``` r
+
+incident_slash_stab <- nyc_pull_dataset("nyc_slash_stab", limit = 3, filters = list(incident_type = "Stabbing"))
+incident_slash_stab
+#> # A tibble: 3 × 4
+#>   incident_id reported_dt         incident_type facility
+#>         <dbl> <dttm>              <chr>         <chr>   
+#> 1      227264 2026-01-24 21:10:00 Stabbing      GRVC    
+#> 2      228684 2026-02-27 22:05:00 Stabbing      RNDC    
+#> 3      228558 2026-02-25 15:43:00 Stabbing      OBCC
+
+# Checking to see the filtering worked
+incident_slash_stab |>
+  distinct(incident_type)
+#> # A tibble: 1 × 1
+#>   incident_type
+#>   <chr>        
+#> 1 Stabbing
+```
+
+## Further investigation.
+
[…]
