### 2026-03-17 1dd39cf
Fixed jaitor issue in utils_request, updated the getting-started vignette, and fixed the name in the data-raw
