+++ vignettes/getting-started.Rmd
@@ -24,13 +24,63 @@ library(dplyr)
 
 ## Introduction
 
-NYC has a population of almost 8.5 million people. By calling 311, residents are able to make comments, inquiries, complaints, and requests to the city agencies. All 311 service requests are contained in the dataset, [found here](https://data.cityofnewyork.us/Social-Services/311-Service-Requests-from-2020-to-Present/erm2-nwe9/about_data). In R, the `nycOpenData` package can be used to pull this data directly.
+Welcome to the `nycOpenData` package, a R package dedicated to helping R users connect to the [NYC Open Data Portal](https://opendata.cityofnewyork.us/)!
 
-The `nycOpenData` package provides a streamlined interface for accessing New York City's vast open data resources. It connects directly to the NYC Open Data Portal. It is currently utilized as a primary tool for teaching data acquisition in [Reproducible Research Using R](https://martinezc1-reproducible-research-using-r.share.connect.posit.cloud/), helping students bridge the gap between raw city APIs and tidy data analysis.
+The `nycOpenData` package provides a streamlined interface for accessing New York City's vast open data resources. It connects directly to the NYC Open Data Portal, helping users bridge the gap between raw city APIs and tidy data analysis. It does this in two ways:
+
+### The `nyc_pull_dataset()` function
+
+The primary way to pull data in this package is the `nyc_pull_dataset()` function, which works in tandem with `nyc_list_datasets()`. You do not need to know anything about API keys or authentication.
+
+The first step would be to call the `nyc_list_datasets()` to see what datasets are in the list and available to use in the `nyc_pull_dataset()` function.
+
+```{r nyc-list-datasets}
+nyc_list_datasets()
+```
+
+The output includes columns such as the dataset title, description, and link to the source. The most important piece is the key, because the key is what's needed for the `nyc_pull_dataset()` function.
+
+```{r nyc-311-pull}
+nyc_311_data <- nyc_pull_dataset(key = "nyc_311", limit = 3)
+```
+
+Now, we have the 3 most recent 311 requests!
+
+### The `nyc_any_dataset()` function
+
+The easiest workflow is to use `nyc_list_datasets()` together with `nyc_pull_dataset()`. However, there are ample datasets on the portal, with new ones being added all the time, and so the list does not have all of the datasets.
+
+In the event that you have a particular dataset you want to use in R that is not in the list, you can use the `nyc_any_dataset()`. The only requirement is the dataset’s API endpoint (a URL provided by the NYC Open Data portal). Here are the steps to get it:
+
+1. On the NYC Open Data Portal, go to the dataset you want to work with.
+2. Click on "Export" (next to the actions button on the right hand side).
+3. Click on "API Endpoint".
+4. Click on "SODA2" for "Version".
+4. Copy the API Endpoint.
+
+Below is an example of how to use the `nyc_any_dataset()` once the API endpoint has been discovered, that will pull the same data as the `nyc_pull_dataset()` example:
+
+```{text}
+nyc_311_data <- nyc_any_dataset(json_link = "https://data.cityofnewyork.us/resource/erm2-nwe9.json", limit = 3)
+```
+
+### Rule of Thumb
+
+While both functions provide access to NYC Open Data, they serve slightly different purposes.
+
+In general:
+- Use `nyc_pull_dataset()` when the dataset is available in `nyc_list_datasets()`
+- Use `nyc_any_dataset()` when working with datasets outside the built-in catalog
+
+Together, these functions allow users to either quickly access curated datasets or flexibly query any dataset available on the NYC Open Data portal.
+
+## Real World Example
+
[…]
