### 2026-03-24 bf0824b
Made significant change to nyc_list_dataset so now it pulls from the nyc open data catalog and not a curated list. Then made changes to nyc_pull_dataset so that it pulled better from this new list. Changed associated documentation
