+++ R/nyc_pull_dataset.R
@@ -1,16 +1,23 @@
-#' Pull a NYC Open Data dataset from the nycOpenData catalog
+#' Pull a NYC Open Data dataset from the NYC Open Data catalog
 #'
-#' Uses a dataset `key` (from `nyc_list_datasets()`) to pull data from NYC Open Data
-#' with sensible defaults (dataset id, default ordering, optional default date field).
+#' Uses a dataset `key` or `uid` from `nyc_list_datasets()` to pull data
+#' from NYC Open Data.
 #'
-#' @param key A dataset key from `nyc_list_datasets()` (e.g., "nyc_311_2010_2019").
+#' Dataset keys are generated from dataset names using
+#' `janitor::make_clean_names()`. Because keys are derived from live catalog
+#' metadata, dataset UIDs are the more stable option.
+#'
+#' @param dataset A dataset key or UID from `nyc_list_datasets()`.
 #' @param limit Number of rows to retrieve (default = 10,000).
 #' @param filters Optional named list of filters. Supports vectors (translated to IN()).
-#' @param date Optional single date (matches all times that day) using the catalog `default_date_field`.
-#' @param from Optional start date (inclusive) using the catalog `default_date_field`.
-#' @param to Optional end date (exclusive) using the catalog `default_date_field`.
-#' @param where Optional raw SoQL WHERE clause. If date/from/to are provided, they are AND-ed with this.
-#' @param order Optional SoQL order override. If NULL, uses catalog `default_order` if present.
+#' @param date Optional single date (matches all times that day) using `date_field`.
+#' @param from Optional start date (inclusive) using `date_field`.
+#' @param to Optional end date (exclusive) using `date_field`.
+#' @param date_field Optional date/datetime column to use with `date`, `from`, or `to`.
+#'   Must be supplied when `date`, `from`, or `to` are used.
+#' @param where Optional raw SoQL WHERE clause. If `date`, `from`, or `to` are
+#'   provided, their conditions are AND-ed with this.
+#' @param order Optional SoQL ORDER BY clause.
 #' @param timeout_sec Request timeout in seconds (default = 30).
 #' @param clean_names Logical; if TRUE, convert column names to snake_case (default = TRUE).
 #' @param coerce_types Logical; if TRUE, attempt light type coercion (default = TRUE).
@@ -19,73 +26,96 @@
 #' @examples
 #' if (interactive() && curl::has_internet()) {
 #'   # Pull by key
-#'   nyc_pull_dataset("nyc_311_2010_2019", limit = 3)
+#'   nyc_pull_dataset("311_service_requests", limit = 3)
+#'
+#'   # Pull by UID
+#'   nyc_pull_dataset("erm2-nwe9", limit = 3)
 #'
 #'   # Filters
-#'   nyc_pull_dataset("nyc_311_2010_2019", limit = 3, filters = list(borough = "QUEENS"))
+#'   nyc_pull_dataset("erm2-nwe9", limit = 3, filters = list(borough = "QUEENS"))
+#'
+#'   # Date filtering
+#'   nyc_pull_dataset(
+#'     "erm2-nwe9",
+#'     from = "2023-01-01",
+#'     to = "2024-01-01",
+#'     date_field = "created_date",
+#'     limit = 100
+#'   )
 #' }
 #' @export
-nyc_pull_dataset <- function(key,
+nyc_pull_dataset <- function(dataset,
                              limit = 10000,
                              filters = list(),
                              date = NULL,
                              from = NULL,
                              to = NULL,
+                             date_field = NULL,
                              where = NULL,
                              order = NULL,
[…]
