+++ R/nyc_list_datasets.R
@@ -1,11 +1,36 @@
 #' List datasets available in nycOpenData
 #'
-#' Returns a catalog of datasets supported by this package.
+#' Retrieves the current NYC Open Data catalog and returns datasets available
+#' for use with `nyc_pull_dataset()`.
 #'
-#' @return A tibble of available datasets.
+#' Keys are generated from dataset names using `janitor::make_clean_names()`.
+#'
+#' @return A tibble of available datasets, including generated `key`, dataset
+#'   `uid`, and dataset `name`.
 #' @examples
-#' nyc_list_datasets()
+#' if (interactive() && curl::has_internet()) {
+#'   nyc_list_datasets()
+#' }
 #' @export
 nyc_list_datasets <- function() {
-  tibble::as_tibble(.nyc_catalog)
+  .nyc_catalog_tbl()
+}
+
+.nyc_catalog_tbl <- function() {
+  raw <- .nyc_dataset_request(
+    dataset_id = "5tqd-u88y",
+    limit = 50000,
+    timeout_sec = 30,
+    clean_names = TRUE,
+    coerce_types = FALSE
+  )
+
+  raw |>
+    dplyr::filter(type == "dataset") |>
+    dplyr::mutate(
+      key = janitor::make_clean_names(name)
+    ) |>
+    dplyr::filter(!is.na(uid), nzchar(uid)) |>
+    dplyr::distinct(uid, .keep_all = TRUE) |>
+    dplyr::relocate(key, uid, name)
 }
