+++ tests/testthat/test-nyc_catalog_and_pull_dataset.R
@@ -1,122 +1,214 @@
 test_that("nyc_list_datasets returns a catalog tibble with expected columns", {
-  cat <- nyc_list_datasets()
-  expect_s3_class(cat, "tbl_df")
-  expect_true(nrow(cat) >= 1)
-
-  expected_cols <- c(
-    "key", "dataset_id", "title", "description",
-    "default_order", "default_date_field",
-    "landing_page_url", "topic"
-  )
-
-  # Allow extra columns; require at least these
-  expect_true(all(expected_cols %in% names(cat)))
-})
-
-test_that("nyc_pull_dataset returns a tibble, respects limits, supports filters + date/from/to (catalog-driven)", {
   skip_if_not_installed("vcr")
   skip_if_not_installed("curl")
+  skip_if_no_cassette("nyc_list_datasets_catalog")
 
-  # pick a stable, known key (update if your catalog key differs)
-  key <- "nyc_311_2010_2019"
+  vcr::use_cassette("nyc_list_datasets_catalog", {
+    cat <- nyc_list_datasets()
+
+    expect_s3_class(cat, "tbl_df")
+    expect_gte(nrow(cat), 1)
+    expect_true(all(c("key", "uid", "name") %in% names(cat)))
+  })
+})
+
+test_that("nyc_pull_dataset returns a tibble, respects limits, supports filters + date/from/to", {
+  skip_if_not_installed("vcr")
+  skip_if_not_installed("curl")
+  skip_if_no_cassette("nyc_pull_dataset_robust")
+
+  dataset_uid <- "uvpi-gqnh"
 
   vcr::use_cassette("nyc_pull_dataset_robust", {
-    # base
-    base <- nyc_pull_dataset(key = key, limit = 2)
+    base <- nyc_pull_dataset(
+      dataset = dataset_uid,
+      limit = 2,
+      timeout_sec = 60
+    )
     expect_s3_class(base, "tbl_df")
-    expect_true(nrow(base) >= 0 && nrow(base) <= 2)
-    expect_true(ncol(base) > 0)
-
-    # r16: snake_case output names (no dots) when clean_names default TRUE
+    expect_gte(nrow(base), 0)
+    expect_lte(nrow(base), 2)
+    expect_gt(ncol(base), 0)
     expect_true(all(!grepl("\\.", names(base))))
 
-    # filters: equality
-    f1 <- nyc_pull_dataset(key = key, limit = 2, filters = list(agency = "NYPD"))
+    f1 <- nyc_pull_dataset(
+      dataset = dataset_uid,
+      limit = 2,
+      filters = list(steward = "None"),
+      timeout_sec = 60
+    )
     expect_s3_class(f1, "tbl_df")
-    expect_true(nrow(f1) >= 0 && nrow(f1) <= 2)
+    expect_gte(nrow(f1), 0)
+    expect_lte(nrow(f1), 2)
 
-    # filters: IN (...)
-    f2 <- nyc_pull_dataset(key = key, limit = 2, filters = list(agency = c("NYPD", "DOT")))
+    f2 <- nyc_pull_dataset(
+      dataset = dataset_uid,
+      limit = 2,
+      filters = list(steward = c("None", "1or2")),
+      timeout_sec = 60
+    )
     expect_s3_class(f2, "tbl_df")
-    expect_true(nrow(f2) >= 0 && nrow(f2) <= 2)
+    expect_gte(nrow(f2), 0)
+    expect_lte(nrow(f2), 2)
 
-    # date branch (only works if catalog default_date_field is set for this key)
[…]
