+++ tests/testthat/helper-vcr.R
@@ -1,21 +1,23 @@
-library(vcr)
-
-# This tells vcr exactly where to find your recordings
-vcr::vcr_configure(
-  dir = "fixtures"
-)
-
 # tests/testthat/helper-vcr.R
 
 if (requireNamespace("vcr", quietly = TRUE)) {
+  vcr::vcr_configure(dir = "fixtures")
+}
 
-  vcr::vcr_configure(
-    # Keep fixtures inside tests/testthat/fixtures
-    dir = file.path("fixtures")
-  )
+if (requireNamespace("webmockr", quietly = TRUE)) {
+  webmockr::enable()
+}
 
-  # Optional but helpful: ensure webmocking is enabled when available
-  if (requireNamespace("webmockr", quietly = TRUE)) {
-    webmockr::enable()
+skip_if_no_cassette <- function(name) {
+  cassette_file <- file.path("fixtures", paste0(name, ".yml"))
+  record_mode <- Sys.getenv("VCR_RECORD", unset = "none")
+
+  if (identical(record_mode, "none") && !file.exists(cassette_file)) {
+    testthat::skip(paste0(
+      "Cassette not recorded yet for '", name,
+      "'. Run with VCR_RECORD=once to create it."
+    ))
   }
 }
+
+options(testthat.snapshot.enabled = FALSE)
