+++ vignettes/analyzing-nyc_wetlands.Rmd
@@ -28,15 +28,15 @@ New York City is home to many wetland features. In an effort to grow awareness o
 
 The `nycOpenData` package provides a streamlined interface for accessing New York City's vast open data resources. It connects directly to the NYC Open Data Portal. It is currently utilized as a primary tool for teaching data acquisition in [Reproducible Research Using R](https://martinezc1-reproducible-research-using-r.share.connect.posit.cloud/), helping students bridge the gap between raw city APIs and tidy data analysis.
 
-By using the `nyc_wetlands()` function, we can gather the most recently listed wetland features in New York City, and filter based upon any of the columns inside the dataset. 
+By using the `p48c-iqtu()` function, we can gather the most recently listed wetland features in New York City, and filter based upon any of the columns inside the dataset. 
 
-> Note: `nyc_wetlands()` automatically sorts in descending order based on the verificationstatusyear column. Due to this order, the first group of rows are `Unverified`, so the verificationstatus year is omitted for those rows.
+> Note: `p48c-iqtu()` automatically sorts in descending order based on the verificationstatusyear column. Due to this order, the first group of rows are `Unverified`, so the verificationstatus year is omitted for those rows.
 
 ## Pulling a Small Sample
 To start, let's pull a small sample to see what the data looks like. By default, the function pulls in the *10,000 most recent* additions, however, let's change that to only see the latest 3 additions. To do this, we can set `limit = 3`.
 
 ```{r small-sample}
-small_sample <- nyc_pull_dataset("nyc_wetlands", limit = 3)
+small_sample <- nyc_pull_dataset("p48c-iqtu", limit = 3)
 small_sample
 
 # Seeing what columns are in the dataset
@@ -49,7 +49,7 @@ Let's now pull the complete dataset to work with:
 
 ## Pulling Full Dataset
 ```{r full-data}
-wetlands_data <- nyc_pull_dataset("nyc_wetlands", limit = 100)
+wetlands_data <- nyc_pull_dataset("p48c-iqtu", limit = 100)
 
 # Let's take a look at what our full dataset looks like
 wetlands_data |>
