### 2026-03-25 #750 @martinezc1
Hello @ronnyhdez !

I hope you are doing well and enjoying the beginning of spring. I understand that you are in the process of finding reviewers, but just wanted to update you regarding a significant update (and improvement) to the package:

1.  Somewhat serendipitously, I discovered that the NYC Open Data Portal has a dataset that has metadata regarding every single dataset on the platform, over 2,000. This is game changing for my package, since I previously created a curated catalog of some of the most popular on the platform, but that only amounted to 40. Of course, I could go 1 by 1 and add even more, but that would not only be extremely tedious, but also not account for any changes to datasets (like names, etc). With that, I updated the nyc_list_datasets() function to pull from the API of the metadata dataset instead of my curated catalog.
2. I made slight adjustments to the nyc_pull_dataset() function due to this. First, I made it so you can pull a dataset either using the key (that is made in the R code for the nyc_list_dataset()) or the uid. Second, I changed some filtering things because originally the curated catalog had a column that represented a date column (if it had one) but now that info is not available. Users can still filter as long as they know what the date column they want to filter is.
3. I changed the getting-started vignette to pull from a dataset that isn’t to big like 311 to make any processing easier.

I totally understand that this may mess with the process, but I actually think this is best case scenario. Now, this is in it’s most complete form and is even more powerful than ever, giving access and info to all datasets.

Let me know if this changes the process, if you have any questions, or anything else. Thank you so much for your time, effort, and patience!
