+++ NAMESPACE
@@ -1,2 +1,6 @@
-exportPattern("^[[:alpha:]]+")
+# Generated by roxygen2: do not edit by hand
+
+export(nyc_any_dataset)
+export(nyc_list_datasets)
+export(nyc_pull_dataset)
 importFrom(rlang,.data)
