### 2026-03-25 #750 @ronnyhdez
Hi @martinezc1!

That's good to hear. I agree with you is better to obtain that "data dictionary" from the site than curating a list manually.

Given that we are still waiting for reviewers response, making this is fine. During the review process it would be better to coordinate this before merging changes to the main branch, but is something we can talk about during the review.

Thanks for explaining the changes!
