+++ index.md
@@ -1,0 +1,198 @@
+# nycOpenData ![nycOpenData logo](reference/figures/nycOpenData_hex.png)
+
+[![CRAN
+status](https://www.r-pkg.org/badges/version/nycOpenData)](https://CRAN.R-project.org/package=nycOpenData)
+[![CRAN
+downloads](https://cranlogs.r-pkg.org/badges/grand-total/nycOpenData?color=blue)](https://r-pkg.org/pkg/nycOpenData)
+[![Lifecycle:
+stable](https://img.shields.io/badge/lifecycle-stable-brightgreen.svg)](https://lifecycle.r-lib.org/articles/stages.html)
+[![Project Status:
+Active](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
+[![Codecov test
+coverage](https://codecov.io/gh/martinezc1/nycOpenData/branch/main/graph/badge.svg)](https://app.codecov.io/gh/martinezc1/nycOpenData)
+[![R-bloggers](https://img.shields.io/badge/Featured%20on-R--bloggers-orange)](https://www.r-bloggers.com/2026/01/nycopendata-a-unified-r-interface-to-nyc-open-data-apis/)
+[![R
+Weekly](https://img.shields.io/badge/Featured%20on-R%20Weekly-blue)](https://rweekly.org/#RintheRealWorld)
+[![R-CMD-check](https://github.com/martinezc1/nycOpenData/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/martinezc1/nycOpenData/actions/workflows/R-CMD-check.yaml)
+
+Version **0.2.1** introduces a streamlined, catalog-driven interface for
+NYC Open Data.
+
+Instead of maintaining dozens of individual dataset wrappers, the
+package now provides three core functions:
+
+- [`nyc_list_datasets()`](https://martinezc1.github.io/nycOpenData/reference/nyc_list_datasets.md)
+  — Browse available datasets from the live NYC Open Data catalog
+- [`nyc_pull_dataset()`](https://martinezc1.github.io/nycOpenData/reference/nyc_pull_dataset.md)
+  — Pull any cataloged dataset by key, with filtering, ordering, and
+  optional date controls
+- [`nyc_any_dataset()`](https://martinezc1.github.io/nycOpenData/reference/nyc_any_dataset.md)
+  — Pull any NYC Open Data dataset directly via its Socrata JSON
+  endpoint
+
+The catalog currently includes 30+ curated NYC Open Data datasets,
+covering topics such as:
+
+- 311 Service Requests
+- For-Hire Vehicles (FHV)
+- Juvenile Justice (rearrest rates + caseloads)
+- School Discharge Reporting
+- Violent & Disruptive School Incidents
+- Detention Admissions
+- Borough/Community District Reports
+- Street Tree Census
+- Urban Park Ranger Animal Condition Responses
+- Permitted Events (Historical)
+- and more
+
+Datasets pulled via
+[`nyc_pull_dataset()`](https://martinezc1.github.io/nycOpenData/reference/nyc_pull_dataset.md)
+automatically apply sensible defaults from the catalog (such as default
+ordering and date fields), while still allowing user control over:
+
+- limit
+- filters
+- date / from / to
+- where
+- order
+- clean_names
+- coerce_types
+
+This redesign reduces maintenance burden, improves extensibility, and
+provides a more scalable interface for working with NYC Open Data.
+
+All functions return clean **tibble** outputs and support filtering
+via  
+`filters = list(field = "value")`.
+
+------------------------------------------------------------------------
+
+## Installation
+
+### From **CRAN**
+
+``` r
+install.packages("nycOpenData")
+```
+
+### Development version (GitHub)
+
+``` r
[…]
