+++ reference/index.md
@@ -1,0 +1,12 @@
+# Package index
+
+## Getting started
+
+Core functions
+
+- [`nyc_list_datasets()`](https://martinezc1.github.io/nycOpenData/reference/nyc_list_datasets.md)
+  : List datasets available in nycOpenData
+- [`nyc_pull_dataset()`](https://martinezc1.github.io/nycOpenData/reference/nyc_pull_dataset.md)
+  : Pull a NYC Open Data dataset from the NYC Open Data catalog
+- [`nyc_any_dataset()`](https://martinezc1.github.io/nycOpenData/reference/nyc_any_dataset.md)
+  : Load Any NYC Open Data Dataset
