+++ articles/analyzing-nyc_events_sealevel.md
@@ -1,0 +1,126 @@
+# Analyzing NYC Climate Projections: Extreme Events and Sea Level Rise
+
+## Introduction
+
+This vignette demonstrates how to use the 38ps-fnsg() function to
+explore projected extreme climate events and sea level rise for New York
+City using the [New York City Climate Projections: Extreme Events and
+Sea Level
+Rise](https://data.cityofnewyork.us/Environment/New-York-City-Climate-Projections-Extreme-Events-a/38ps-fnsg/about_data)
+dataset on the NYC Open Data portal.
+
+The dataset provides projections under different climate scenarios,
+including:
+
+- Number of heatwaves per year
+- Cooling and heating degree days
+- Projected sea level rise
+
+Researchers, city planners, and policymakers can use this information to
+understand future climate risks, prepare for extreme weather events, and
+plan adaptation strategies.
+
+## Retrieve a Sample of Data
+
+``` r
+sample_data <- nyc_pull_dataset("38ps-fnsg", limit = 10)
+nyc_list_datasets()
+#> # A tibble: 2,389 × 26
+#>    key             uid   name  datasetinformation_a…¹ description type  category
+#>    <chr>           <chr> <chr> <chr>                  <chr>       <chr> <chr>   
+#>  1 orr_points      dhcf… ORR_… Mayor's Office of Cli… "Completed… data… City Go…
+#>  2 off_year_and_s… ups9… Off-… Campaign Finance Boar… "This data… data… City Go…
+#>  3 parks_trails    vjbm… Park… Department of Parks a… "Location … data… Environ…
+#>  4 where_the_doll… qhm5… Wher… Mayor's Office of Man… "This data… data… City Go…
+#>  5 citywide_mobil… rb38… City… Department of Transpo… "In 2020, … data… Transpo…
+#>  6 x2013_2014_sch… ac4n… 2013… Department of Educati… "School Lo… data… Educati…
+#>  7 tourism_grants  rma9… Tour… Brooklyn Borough Pres… "Final lis… data… City Go…
+#>  8 citywide_cash_… 9jbx… City… Human Resources Admin… "Total num… data… Social …
+#>  9 nyc_domain_reg… ymvu… .nyc… Office of Technology … ".nyc doma… data… Business
+#> 10 nycha_developm… evjd… NYCH… New York City Housing… "Contains … data… Housing…
+#> # ℹ 2,379 more rows
+#> # ℹ abbreviated name: ¹​datasetinformation_agency
+#> # ℹ 19 more variables: legislativecompliance_datasetfromtheopendataplan <chr>,
+#> #   url <chr>, update_datemadepublic <chr>, update_updatefrequency <chr>,
+#> #   last_data_updated_date <chr>,
+#> #   legislativecompliance_candatasetfeasiblybeautomated <chr>,
+#> #   update_automation <chr>, legislativecompliance_hasdatadictionary <chr>, …
+
+sample_data
+#> # A tibble: 10 × 16
+#>    period           sea_lelel_rise number_of_days_year_…¹ number_of_days_year_…²
+#>    <chr>            <chr>                           <dbl>                  <dbl>
+#>  1 Baseline (1981-… n/a                                69                     17
+#>  2 2030s (10th Per… 6 in                               85                     27
+#>  3 2030s (25th Per… 7 in                               85                     27
+#>  4 2030s (75th Per… 11 in                              99                     46
+#>  5 2030s (90th Per… 13 in                             104                     54
+#>  6 2050s (10th Per… 12 in                              91                     32
+#>  7 2050s (25th Per… 14 in                              99                     38
+#>  8 2050s (75th Per… 19 in                             100                     62
+#>  9 2050s (90th Per… 23 in                             121                     69
[…]
