+++ authors.md
@@ -1,0 +1,51 @@
+# Authors and Citation
+
+## Authors
+
+- **Christian Martinez**. Author, maintainer.
+  [](https://orcid.org/0009-0005-6026-6454)  
+  martinezc1
+
+- **Crystal Adote**. Contributor.  
+  crystalna20
+
+- **Jonah Dratfield**. Contributor.  
+  jdratfield38
+
+- **Joyce Escatel-Flores**. Contributor.  
+  JoyceEscatel
+
+- **Rob Hutto**. Contributor.  
+  robhutto
+
+- **Isley Jean-Pierre**. Contributor.  
+  ijpier
+
+- **Shannon Joyce**. Contributor.  
+  shannonjoyce
+
+- **Laura Rose-Werner**. Contributor.  
+  laurarosewerner
+
+- **Emma Tupone**. Contributor.  
+  emmatup0205
+
+- **Xinru Wang**. Contributor.  
+  xrwangxr
+
+## Citation
+
+Source:
+[`inst/CITATION`](https://github.com/martinezc1/nycOpenData/blob/main/inst/CITATION)
+
+Martinez C (2026). *nycOpenData: Accessing and Analyzing NYC Open Data
+in R*. R package version 0.2.1,
+<https://github.com/martinezc1/nycOpenData>.
+
+    @Manual{,
+      title = {nycOpenData: Accessing and Analyzing NYC Open Data in R},
+      author = {Christian Martinez},
+      year = {2026},
+      note = {R package version 0.2.1},
+      url = {https://github.com/martinezc1/nycOpenData},
+    }

[…]
