### 2026-04-13 #750 @ronnyhdez
Hi @martinezc1,

We now have two reviewers on board for the package. Thanks @donghl17 and @MichaelPascale for your time!

To keep things organized, I suggest holding off on further changes until the reviews are complete. Reviewers can work from this commit:
https://github.com/martinezc1/nycOpenData/tree/cb213e4695243c419ac869395d6c962a9dba4237

@martinezc1 If you need to implement changes during the review, please consider grouping them in a new release or a clearly defined set of commits. This helps reviewers track how feedback is addressed and avoids mixing review related changes with other ongoing work.
