### 2026-05-03 #750 @maurolepore
I see @ronnyhdez has guided this through editor checks and assigned @donghl17 and @MichaelPascale as reviewers. I see reviews are due soon. Looking forward to seeing this move through review.

Thanks for sharing your work with the rOpenSci community.
