### 2026-05-04 #750 @MichaelPascale
I have no previous relationship to the package authors.

Restate the specific problems this package solves and add to the README (from what you've posted in this thread above). The README description describes `nycOpenData` as supporting data literacy (admirably) but does not state plainly that the purpose of the package is to list and download tables from the city's Socrata API.

Installation was straightforward. Perhaps alert the development user that webmockr is used. Alert the development user of this.

Vignettes run but are largely of redundant structure. It is not clear that the many separate vignettes are necessary to describe the three functions.

Please describe functions in more detail. Though `nyc_list_datasets()` has a very simple job, its documentation does not really provide transparency to the naive user into how it accomplishes its job. I might suggest that you mention the actual portal it pulls from.

The returned tibble includes both a UID and a key, as well as several undocumented columns. Explain the columns returned more thoroughly. If users have to refer to the Socrata data portal to use this package in a classroom setting, then still they might as well have interfaced directly with the API themselves.

Briefly describe concepts necessary to navigate the NYC Open Datasets API, including the difference between the UID and key. Update this in the README too.

Please provide more examples of what you might do with each function. The `nyc_list_datasets()` help just shows calling the function, which is no more than the usage section has. You might show how to sort and locate a particular datset key, for instance.

Improved documentation will help this package better fulfill its educational aims.

---

Congratulations authors on your package! It is lovely to see this collaboration among students and instructors resulting in a product. I hope that the following feedback will be helpful in continuing to improve `nycOpenData`.

Reviewer Summary: `nycOpenData` is intended as an educational tool for students to easily interact with the NYC Open Datasets Socrata API in R with little overhead and with little prerequisite technical knowledge. It is a small package, providing three methods to make simple queries. It's aims could be clarified slightly, and improved function documentation would be the biggest step the authors could take toward fulfilling these aims. The package does indeed provide easy access to the NYC Open Datasets tables, but it may not add sufficient functionality on its own to warrant support. The current version is available on CRAN.

The following issues should be addressed, either fixed or documented more clearly.

- On using the coerce_types argument, the internal function `.parse_iso_datetime_safe()` will drop any timezone offsets. I imagine most New York data has either UTC or EDT/EST timezones. This information should not be dropped, unless you can guarantee that all the timestamps are already UTC.
- The borough population vignette references a `xywu-7bv9()` "function", which is either incorrect or worded misleadingly. Others have similar issues, and some read as adaptations of homework assignments that could now be merged into one example vignette for the purposes of packaging. Perhaps these are holdovers from previous versions of the package, which, to my understanding, had separate functions for different tables. In any case, they should be removed from the submitted package.
- - Include in the documentation the information about the catalog table implicitly fetched by `nyc_list_datasets()` (i.e. '5tqd-u88y').
- Check the version strings are updated consistently throughout the codebase and documentation.

In addition to the changes described above, the following recommendations are offered:

- The Socrata HTTP response offers the X-SODA2-Types headers. Consider adding an option (default) to these for type inference and coercion rather than guessing.
- Consider caching the table retrieved by .nyc_catalog_tbl, which is run on every pull and call to nyc_list_datasets. If this is used by students, imagine that perhaps there will be many interactive calls to this function as students troubleshoot their code.
- Consider allowing the user to batch their downloads by providing a batch/batchsize argument for `nyc_pull_dataset()` instead of the current default `limit`.
- There are references to Socrata and SoQL that are provided without context. Briefly explain these in the documentation or provide links.

Finally, the following are offered as tips, and are smaller work than the recommendations above.
- The vignettes are well done, but there is some redundancy across the vignettes. Most of them essentially show how to do `nyc_pull_dataset()` once on a particular dataset, and then the rest is use of `dplyr` and `ggplot2` primarily.
- Remove .nyc_abort or utilize it consistently. Only some errors use this.
- Consider adding a clear disclaimer that this tool is not maintained by the City of New York.
- It would seem important to describe how this package will be kept up to date with respect to changes to the NYC Open Datasets API.
- Explain the filters supported by `nyc_pull_dataset()` with a description, in addition to the examples.
- In the README where it says "while still allowing user control over:...", write out a description rather than listing arguments of the function.
- A citation file could be added (each vignette ends with a citation, so place it in CITATION per https://devguide.ropensci.org/pkg_building.html#citation-file).

Note that alternatives to this package exist, including direct download from the portal site, use of `RSocrata`, direct query with `httr`, or e.g. `jsonlite`:

    # With RSocrata...
    read.socrata("https://data.cityofnewyork.us/resource/5tqd-u88y.json?%24limit=50000")

    # Or even more bare-bones...

    tibble(read_json("https://data.cityofnewyork.us/resource/5tqd-u88y.json?%24limit=50000", simplifyVector=T))

RSocrata, which is created by the City of Chicago, appears to have little recent activity (though not necessarily unmaintained). Though it is open source and encourages public contribution on GitHub, it does not appear under a free license. Perhaps mention these as additional rationale for the project as best-in-category. Emphasize that the smaller size of this project will ensure it is more maintainable. Perhaps also consider reaching out to the RSocrata authors to determine the state of that project.

I can see from the history that the package was originally organized as many separate functions for datasets, and I want to let you know that many times I myself have started to create a package --- and learned a whole lot along the way --- to eventually find that what I had set out to do could be done much more simply or with a tool I was not aware of. Consider whether this package is likely to become a dependency of other projects, or will rather be used solely in the educational context. If it starts to seem like the package is not necessary to fulfill the aims, then perhaps consider making these available as a series of blog posts or adapting them into open course materials (the GitHub pages site is already quite nice!).

`nycOpenData` is valuable (in my perspective) in that a new user could potentially figure out how to load a dataset, e.g., in the first five minutes of class time, fairly frictionlessly. If then it is to be an educational package, place emphasis on these advantages and do continue developing them further. Are there analyses such a package might support, for example? Consider whether a user will be likely to go to the data portal website regardless of whether or not they use this package.
