### 2026-05-08 785e5bb
Made significant changes based on rOpenSci reviewers
