+++ paper.bib
@@ -1,39 +1,0 @@
-@Manual{r-core,
-  title = {R: A Language and Environment for Statistical Computing},
-  author = {{R Core Team}},
-  organization = {R Foundation for Statistical Computing},
-  year = {2024},
-  url = {https://www.R-project.org/},
-}
-
-@book{reproducible_research,
-  title = {Reproducible Research Using R},
-  author = {Martinez, Christian},
-  year = {2024},
-  publisher = {Self-Published},
-  note = {Primary pedagogical tool for NYC data acquisition curriculum},
-}
-
-@misc{soda_api,
-  author = {Socrata},
-  title = {Open Data Network: Socrata Open Data API (SODA)},
-  year = {2026},
-  url = {https://dev.socrata.com/},
-}
-
-@Manual{jsonlite,
-  title = {jsonlite: A Robust, High-Speed JSON Parser and Generator for R},
-  author = {Jeroen Ooms},
-  year = {2023},
-  url = {https://CRAN.R-project.org/package=jsonlite},
-}
-
-@Article{testthat,
-  author = {Hadley Wickham},
-  title = {testthat: Get Started with Testing},
-  journal = {The R Journal},
-  year = {2011},
-  volume = {3},
-  pages = {5--10},
-  url = {https://journal.r-project.org/archive/2011-1/RProject_2011-1_Wickham.pdf},
-}
