+++ _pkgdown.yml
@@ -8,14 +8,6 @@ articles:
   - title: "Vignettes"
     contents:
       - getting-started
-      - Using-Local-Law18-Pay-Report
-      - analyzing-nyc-borough-population
-      - Analyzing-the-Local-Law-97-of-2021-Pets-in-Shelter-Report
-      - analyzing-nyc_events_sealevel
-      - analyzing-nyc_wetlands
-      - analyzing-nyc-shooting-incidents
-      - nyc_ha_violations
-      - investigating-inmate-incidents
 reference:
   - title: "Getting started"
     desc: "Core functions"
