+++ tests/testthat/fixtures/nyc_any_dataset_robust.yml
@@ -1,58 +1,58 @@
 http_interactions:
 - request:
     method: GET
-    uri: https://data.cityofnewyork.us/resource/erm2-nwe9.json?%24limit=2
+    uri: https://data.cityofnewyork.us/resource/xywu-7bv9.json?%24limit=2
   response:
     status: 200
     headers:
       server: nginx
-      date: Wed, 04 Mar 2026 16:28:11 GMT
+      date: Fri, 08 May 2026 13:23:50 GMT
       content-type: application/json;charset=utf-8
       transfer-encoding: chunked
       connection: keep-alive
       access-control-allow-origin: '*'
-      etag: '"Zm94dHJvdC4xNTY1NV8xODNfMTkwNjQ3OTVobG5fTGhXQXVfNHU0VFJobmllWmtCLUh3---gziptJLR-gMf_Bsz_mD3dTS7JfWPHGU--gzip--gzip"'
-      x-soda2-fields: '["unique_key","created_date","closed_date","agency","agency_name","complaint_type","descriptor","descriptor_2","location_type","incident_zip","incident_address","street_name","cross_street_1","cross_street_2","intersection_street_1","intersection_street_2","address_type","city","landmark","facility_type","status","due_date","resolution_description","resolution_action_updated_date","community_board","council_district","police_precinct","bbl","borough","x_coordinate_state_plane","y_coordinate_state_plane","open_data_channel_type","park_facility_name","park_borough","vehicle_type","taxi_company_borough","taxi_pick_up_location","bridge_highway_name","bridge_highway_direction","road_ramp","bridge_highway_segment","latitude","longitude","location",":@computed_region_f5dn_yrer",":@computed_region_yeji_bk3q",":@computed_region_sbqj_enih",":@computed_region_92fq_4b7q"]'
-      x-soda2-types: '["text","floating_timestamp","floating_timestamp","text","text","text","text","text","text","text","text","text","text","text","text","text","text","text","text","text","text","floating_timestamp","text","floating_timestamp","text","text","text","text","text","number","number","text","text","text","text","text","text","text","text","text","text","number","number","point","number","number","number","number"]'
+      etag: '"Zm94dHJvdC4xNjM0MF8yXzc1UzBucHJSMl9DLWZ0Z3dFdnhQak91MHBWR1dJ---gzipcvckzktN_gy8OcIPVW7cfYjm5Yg--gzip--gzip"'
+      x-soda2-fields: '["age_group","borough","_1950","_1950_boro_share_of_nyc_total","_1960","_1960_boro_share_of_nyc_total","_1970","_1970_boro_share_of_nyc_total","_1980","_1980_boro_share_of_nyc_total","_1990","_1990_boro_share_of_nyc_total","_2000","_2000_boro_share_of_nyc_total","_2010","_2010_boro_share_of_nyc_total","_2020","_2020_boro_share_of_nyc_total","_2030","_2030_boro_share_of_nyc_total","_2040","_2040_boro_share_of_nyc_total"]'
+      x-soda2-types: '["text","text","number","number","number","number","number","number","number","number","number","number","number","number","number","number","number","number","number","number","number","number"]'
       x-soda2-data-out-of-date: 'false'
-      x-soda2-truth-last-modified: Wed, 04 Mar 2026 02:33:48 GMT
-      x-soda2-secondary-last-modified: Wed, 04 Mar 2026 02:33:48 GMT
-      last-modified: Wed, 04 Mar 2026 02:33:48 GMT
[…]
