+++ articles/getting-started.md
@@ -1,0 +1,314 @@
+# Getting Started with nycOpenData
+
+``` r
+knitr::opts_chunk$set(warning = FALSE, message = FALSE)
+library(nycOpenData)
+library(ggplot2)
+library(dplyr)
+#> 
+#> Attaching package: 'dplyr'
+#> The following objects are masked from 'package:stats':
+#> 
+#>     filter, lag
+#> The following objects are masked from 'package:base':
+#> 
+#>     intersect, setdiff, setequal, union
+```
+
+## Introduction
+
+Welcome to the `nycOpenData` package, a R package dedicated to helping R
+users connect to the [NYC Open Data
+Portal](https://opendata.cityofnewyork.us/)!
+
+The `nycOpenData` package provides a streamlined interface for accessing
+New York City’s vast open data resources. It connects directly to the
+NYC Open Data Portal, helping users bridge the gap between raw city APIs
+and tidy data analysis. It does this in two ways:
+
+### The `nyc_pull_dataset()` function
+
+The primary way to pull data in this package is the
+[`nyc_pull_dataset()`](https://martinezc1.github.io/nycOpenData/reference/nyc_pull_dataset.md)
+function, which works in tandem with
+[`nyc_list_datasets()`](https://martinezc1.github.io/nycOpenData/reference/nyc_list_datasets.md).
+Most workflows do not require users to manually construct API requests
+or manage authentication details.
+
+The first step would be to call the
+[`nyc_list_datasets()`](https://martinezc1.github.io/nycOpenData/reference/nyc_list_datasets.md)
+to see what datasets are in the list and available to use in the
+[`nyc_pull_dataset()`](https://martinezc1.github.io/nycOpenData/reference/nyc_pull_dataset.md)
+function. This provides information about datasets available through the
+live NYC Open Data catalog used by the package.
+
+``` r
+catalog <- nyc_list_datasets()
+
+catalog |>
+  filter(grepl("collision", name, ignore.case = TRUE)) |>
+  select(key, uid, name)
+#> # A tibble: 4 × 3
+#>   key                                                                uid   name 
+#>   <chr>                                                              <chr> <chr>
+#> 1 collisions_involving_vehicles_managed_by_department_of_health_and… knr6… Coll…
+#> 2 motor_vehicle_collisions_vehicles                                  bm4k… Moto…
+#> 3 motor_vehicle_collisions_person                                    f55k… Moto…
+#> 4 motor_vehicle_collisions_crashes                                   h9gi… Moto…
+```
+
+The output includes columns such as the dataset title, description, and
+link to the source. The most important pieces are the key **and** uid.
+You need **either** in order to use the
+[`nyc_pull_dataset()`](https://martinezc1.github.io/nycOpenData/reference/nyc_pull_dataset.md)
+function. You can put **either** the key value or uid value into the
+`dataset =` filter inside of
+[`nyc_pull_dataset()`](https://martinezc1.github.io/nycOpenData/reference/nyc_pull_dataset.md).
+
+For instance, if we want to pull the the dataset
+`Motor Vehicle Collisions - Crashes`, we can use either of the methods
+below:
+
+``` r
+nyc_motor_vehicle_collisions_data <- nyc_pull_dataset(
+  dataset = "h9gi-nx95", limit = 2)
+
+nyc_motor_vehicle_collisions_data <- nyc_pull_dataset(
+  dataset = "motor_vehicle_collisions_crashes", limit = 2)
+```
+
[…]
