+++ articles/index.md
@@ -1,0 +1,6 @@
+# Articles
+
+### Vignettes
+
+- [Getting Started with
+  nycOpenData](https://martinezc1.github.io/nycOpenData/articles/getting-started.md):
