+++ pkgdown.yml
@@ -1,0 +1,9 @@
+pandoc: 3.1.3
+pkgdown: 2.2.0
+pkgdown_sha: ~
+articles:
+  getting-started: getting-started.html
+last_built: 2026-05-08T13:32Z
+urls:
+  reference: https://martinezc1.github.io/nycOpenData/reference
+  article: https://martinezc1.github.io/nycOpenData/articles

[…]
